requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
django-filter==21.1
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.3
iniconfig==1.1.1
isort==5.9.3
mccabe==0.6.1
packaging==21.0
pep8-naming==0.12.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.1
requests==2.26.0
testfixtures==6.18.3
toml==0.10.2
urllib3==1.26.7
asgiref==3.2.10
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1

